+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 Changelog
 =========
 
+## 6.21.0 (2022-07-20)
+
+### Enhancements
+
+* Add `configuration.reportBackgroundAppHangs` to allow background app hangs to be reported.
+  [#1439](https://github.com/bugsnag/bugsnag-cocoa/pull/1439)
+
+* Add `freeMemory`, `memoryLimit` and `memoryUsage` to `metaData.app`.
+  Always report the device (not app) free memory in `device.freeMemory`.
+  [#1435](https://github.com/bugsnag/bugsnag-cocoa/pull/1435)
+
 ## 6.20.0 (2022-07-13)
 
 ### Enhancements